+++ src/main/java/web/service/UserService.java
@@ -10,10 +10,10 @@ public interface UserService {
 
     void saveUser(User user);
 
-    User showUser (int id);
+    User showUser (Long id);
 
-    void update(int id, User user);
+    void update(Long id, User user);
 
-    void delete(int id);
+    void delete(Long id);
 }
